refactor(chat): split chat-message.tsx into per-role files
Extracts UserMessage, AssistantMessage, and useThinkingTimer out of the
343-line chat-message.tsx monolith into focused single-responsibility files.
chat-message.tsx is reduced to a ~38-line dispatcher.
